fix(backend): CSV download retrieves all results (#1362)
* Move file

* Fix compiler

* Update comments
